requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
face-recognition==1.3.0
opencv-python-headless==4.9.0.80
numpy==1.24.3
Pillow==10.2.0
